better DI and Xor mappings
