+++ test/uk/gov/hmrc/decisionservice/DecisionServiceSpec.scala
@@ -1,10 +1,20 @@
 package uk.gov.hmrc.decisionservice
 
 import uk.gov.hmrc.decisionservice.model.rules.{>>>, Facts}
-import uk.gov.hmrc.decisionservice.service.DecisionServiceInstance
+import uk.gov.hmrc.decisionservice.ruleengine.{RuleEngineInstance, RulesFileMetaData}
+import uk.gov.hmrc.decisionservice.service.DecisionService
 import uk.gov.hmrc.play.test.UnitSpec
 
 class DecisionServiceSpec extends UnitSpec {
+
+  object DecisionServiceTestInstance extends DecisionService {
+    lazy val maybeSectionRules = loadSectionRules()
+    val csvSectionMetadata = List(
+      (7, 3, "/business_structure.csv", "BusinessStructure"),
+      (9, 2, "/personal_service.csv", "PersonalService"),
+      (3, 2, "/matrix.csv", "Matrix")
+    ).collect{case (q,r,f,n) => RulesFileMetaData(q,r,f,n)}
+  }
 
   "decision service" should {
     "produce correct decision for a sample fact set leading to section exit" in {
@@ -28,7 +38,7 @@ class DecisionServiceSpec extends UnitSpec {
         "10" -> >>>("yes"))
       )
 
-      val maybeDecision = facts ==>: DecisionServiceInstance
+      val maybeDecision = facts ==>: DecisionServiceTestInstance
       maybeDecision.isRight shouldBe true
       maybeDecision.map { decision =>
         decision.value shouldBe "exit - out of IR35"
@@ -55,7 +65,7 @@ class DecisionServiceSpec extends UnitSpec {
         "10" -> >>>("yes"))
       )
 
-      val maybeDecision = facts ==>: DecisionServiceInstance
+      val maybeDecision = facts ==>: DecisionServiceTestInstance
       maybeDecision.isRight shouldBe true
       maybeDecision.map { decision =>
         decision.value shouldBe "specialCase"
